Refactor event-specific functions into separate php partial
